requirements.txt
# AGORA AI Backend Requirements

# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# 数据库
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# 认证
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# 环境变量
python-dotenv==1.0.0

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis (可选)
redis==5.0.1

# WebSocket
websockets==12.0

# 工具库
python-dateutil==2.8.2
pytz==2023.3

# 日志
loguru==0.7.2

# 测试 (开发环境)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# 代码质量 (开发环境)
black==23.12.1
flake8==7.0.0
mypy==1.8.0
